🐛 fix(email): address PR review — stdin, recipients, attachment size
- Error immediately when no body source and stdin is a TTY (was blocking)
- Validate at least one recipient before composing SMTP message
- Use int64 for AttachmentInfo.Size (was uint32, could truncate)
- Remove dead filepath.Separator replace after filepath.Base

+++ internal/email/message.go
@@ -17,7 +17,7 @@ type Envelope struct {
 // AttachmentInfo holds metadata about a message attachment (no content).
 type AttachmentInfo struct {
 	Filename string `json:"filename"`
-	Size     uint32 `json:"size"`
+	Size     int64  `json:"size"`
 	MIMEType string `json:"mime_type"`
 }
